pyproject.toml
[project]
name = "contract-review-env"
version = "0.1.0"
description = "OpenEnv environment for AI-powered legal contract clause review"
readme = "README.md"
requires-python = ">=3.10"
license = "BSD-3-Clause"
dependencies = [
    "openenv-core[core]>=0.2.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.0.0",
    "scikit-learn>=1.3.0",
]

[project.scripts]
server = "contract_review_env.server.app:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["contract_review_env*", "server*", "models*", "scenarios*", "graders*"]
